Fix virtual env install
The setup_env script doesn't work as intended with the default values of
using a python virtual environment.

If `pip` is not installed on the host, then it will fail the process.

The issue is that the `case` where the bash script is meant to set the
`PIP` binary to be used is incorrectly written, so it will not go into
any of the 2 options.

Removing the ' from the values options in the `case` resolves the
problem.