Fix python-lxml dependencies
Task rhol_crc reports the following error:
  "The `lxml` module is not importable. Check the requirements."

Something similar happens on the libvirt_manager.

This is because the destination host is missing the `python-lxml`
dependency, when we can see that it is actually used in the code [1].

There is a PR in the module to update the docs [2], though I disagree
with it, because I believe it should be `python-lxml` instead of `lxml`,
like in the `virt_net` module of the same `community.libvirt`
collection.

This patch adds the `python-lxml` dependency to fix this issue.

[1]: https://github.com/ansible-collections/community.libvirt/blob/4f7fa5cc5b3a28a5f5fa44a30a45885f9ed88e22/plugins/modules/virt.py#L160
[2]: https://github.com/ansible-collections/community.libvirt/pull/166
[3]: https://docs.ansible.com/ansible/latest/collections/community/libvirt/virt_net_module.html#requirements